Try incorporate more Variable usage
You have the layout down and have replicated the transcript!  The one problem is that while your print statements are great for this transcript, if the student's info or grades were to change then you would have to rewrite a lot of code.

I've suggested a few changes to replace static strings with variables instead, to make your program more robust and dynamic, so that if the student's name were change, we can rely on variables instead of only static print statements.

See if you can create more variables to store information and substitute those into your print statements!